tutorials/05_채팅UI_만들기/requirements.txt
streamlit>=1.32.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0

